Refactored UI components: - Replaced QR, Camera, and More buttons with a Refresh button - Implemented real-time filter count updates for leads and appointments - Added sorting by modified time (most recent first) - Improved subtitle formatting for date and time - Integrated human-readable 'last updated' timestamps (e.g., '2 hours ago')
