+++ sahayog/sahayog/workspace/sahayog_crm/sahayog_crm.json
@@ -1,0 +1,38 @@
+{
+ "charts": [],
+ "content": "[{\"id\":\"FCWVyAYYX8\",\"type\":\"custom_block\",\"data\":{\"custom_block_name\":\"Sahayog CRM Dashboard\",\"col\":12}}]",
+ "creation": "2025-04-23 14:50:14.614663",
+ "custom_blocks": [
+  {
+   "custom_block_name": "Sahayog CRM Dashboard",
+   "label": "Sahayog CRM Dashboard"
+  }
+ ],
+ "docstatus": 0,
+ "doctype": "Workspace",
+ "for_user": "",
+ "hide_custom": 0,
+ "icon": "crm",
+ "idx": 0,
+ "indicator_color": "green",
+ "is_hidden": 0,
+ "label": "Sahayog CRM",
+ "links": [],
+ "modified": "2025-06-12 15:55:02.406466",
+ "modified_by": "Administrator",
+ "module": "Sahayog",
+ "name": "Sahayog CRM",
+ "number_cards": [],
+ "owner": "Administrator",
+ "parent_page": "",
+ "public": 1,
+ "quick_lists": [],
+ "roles": [
+  {
+   "role": "Sales User"
+  }
+ ],
+ "sequence_id": 68.0,
+ "shortcuts": [],
+ "title": "Sahayog CRM"
+}
